Explicitly cast usages of `irq_t` types
On some platforms `irq_t` is not defined as an `int` resulting in the `userError`
messages failing to type check. This adds an explicit cast

+++ src/object/interrupt.c
@@ -55,14 +55,14 @@ decodeIRQControlInvocation(word_t invLabel, word_t length,
 
         if (isIRQActive(irq)) {
             current_syscall_error.type = seL4_RevokeFirst;
-            userError("Rejecting request for IRQ %u. Already active.", irq);
+            userError("Rejecting request for IRQ %u. Already active.", (int)irq);
             return EXCEPTION_SYSCALL_ERROR;
         }
 
         lu_ret = lookupTargetSlot(cnodeCap, index, depth);
         if (lu_ret.status != EXCEPTION_NONE) {
             userError("Target slot for new IRQ Handler cap invalid: cap %lu, IRQ %u.",
-                      getExtraCPtr(buffer, 0), irq);
+                      getExtraCPtr(buffer, 0), (int)irq);
             return lu_ret.status;
         }
         destSlot = lu_ret.slot;
@@ -70,7 +70,7 @@ decodeIRQControlInvocation(word_t invLabel, word_t length,
         status = ensureEmptySlot(destSlot);
         if (status != EXCEPTION_NONE) {
             userError("Target slot for new IRQ Handler cap not empty: cap %lu, IRQ %u.",
-                      getExtraCPtr(buffer, 0), irq);
+                      getExtraCPtr(buffer, 0), (int)irq);
             return status;
         }
